fix(slack): remove control-plane canvas
