+++ slack-bridge/pinet-home-tabs.ts
@@ -1,5 +1,5 @@
 import type { ExtensionContext } from "@mariozechner/pi-coding-agent";
-import type { BrokerControlPlaneDashboardSnapshot } from "./broker/control-plane-canvas.js";
+import type { BrokerControlPlaneDashboardSnapshot } from "./broker/control-plane-dashboard.js";
 import { probeGitBranch } from "./git-metadata.js";
 import {
   publishSlackHomeTab,
